+++ core/frontend/src/Viewport.ts
@@ -14,7 +14,7 @@ import { BeCursor } from "./tools/Tool";
 import { EventController } from "./tools/EventController";
 import { AuxCoordSystemState, ACSDisplayOptions } from "./AuxCoordSys";
 import { IModelConnection } from "./IModelConnection";
-import { HitDetail } from "./HitDetail";
+import { HitDetail, SnapDetail } from "./HitDetail";
 import { DecorateContext, SceneContext } from "./ViewContext";
 import { TileRequests } from "./tile/TileTree";
 import { LegacyMath } from "@bentley/imodeljs-common/lib/LegacyMath";
@@ -921,7 +921,7 @@ export class Viewport {
   public get pixelsPerInch() { /* ###TODO: This is apparently unobtainable information in a browser... */ return 96; }
   public get viewCmdTargetCenter(): Point3d | undefined { return this._viewCmdTargetCenter; }
   public set viewCmdTargetCenter(center: Point3d | undefined) { this._viewCmdTargetCenter = center ? center.clone() : undefined; }
-  public get backgroundMapPlane() { return this.view.displayStyle.getBackgroundMapPlane(); }
+  public get backgroundMapPlane() { return this.view.displayStyle.backgroundMapPlane; }
 
   /**
    * Sets a function which can customize the appearance of features within a viewport.
@@ -1103,6 +1103,7 @@ export class Viewport {
     const vf = ViewFrustum.createFromViewport(this, view);
     if (undefined === vf)
       return ViewStatus.InvalidViewport;
+
     this._viewFrustum = vf;
 
     this.sync.invalidateRenderPlan();
@@ -1560,7 +1561,7 @@ export class Viewport {
     if (!context.viewport.view.is3d())
       return; // Not valuable feedback in 2d...
 
-    if (!hit.isSnapDetail() || !hit.normal || hit.isPointAdjusted())
+    if (!(hit instanceof SnapDetail) || !hit.normal || hit.isPointAdjusted)
       return; // AccuSnap will flash edge/segment geometry if not a surface hit or snap location has been adjusted...
 
     const graphic = context.createWorldOverlay();