+++ core/backend/CHANGELOG.md
@@ -1,6 +1,31 @@
 # Change Log - @bentley/imodeljs-backend
 
-This log was last generated on Fri, 10 Aug 2018 05:06:20 GMT and should not be manually modified.
+This log was last generated on Wed, 15 Aug 2018 17:08:54 GMT and should not be manually modified.
+
+## 0.117.0
+Wed, 15 Aug 2018 17:08:53 GMT
+
+### Updates
+
+- Flesh out DisplayStyle API; Support iteration of ViewDefinitions
+
+## 0.116.0
+Wed, 15 Aug 2018 15:13:19 GMT
+
+*Version update only*
+
+## 0.115.0
+Tue, 14 Aug 2018 15:21:27 GMT
+
+*Version update only*
+
+## 0.114.0
+Tue, 14 Aug 2018 12:04:18 GMT
+
+### Updates
+
+- Use binary transfer for loadNativeAsset
+- added IModelDb.Views.getThumbnail
 
 ## 0.113.0
 Fri, 10 Aug 2018 05:06:20 GMT